requirements.txt
alembic==1.7.6
aniso8601==9.0.1
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
Django==3.2.9
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-WTF==1.0.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
Mako==1.1.6
MarkupSafe==2.0.1
mongoengine==0.23.1
outcome==1.1.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.3
pyOpenSSL==21.0.0
pyrsistent==0.18.1
pytz==2021.3
requests==2.27.1
rsa==4.8
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
telebot==0.0.4
Telethon==1.24.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
webdriver-manager==3.5.2
Werkzeug==2.0.3
wsproto==1.0.0
WTForms==3.0.1
